Before tidying up

+++ computeCoordinates.cpp
@@ -23,8 +23,8 @@ void wet::computecoordinates(int k)
 	if(rank==ROOT)
 	{
 		kTot=k+rank*(k2-(Lx%size+1)*k1)-k1;
-		xk=int(kTot/(float) (Lz*Ly));		//Index node of x coordinate 
-		yk=int((kTot-xk*Lz*Ly)/(float) Lz);	//Index node of y coordinate
+		xk=kTot/(Lz*Ly);		//Index node of x coordinate 
+		yk=(kTot-xk*Lz*Ly)/Lz;	//Index node of y coordinate
 		zk=kTot-xk*Lz*Ly-yk*Lz;			//Index node of z coordinate
 	}
 	if(rank>ROOT)
